% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_taxon.R
\name{format_name}
\alias{format_name}
\title{Formatting scientific names}
\usage{
format_name(
  df,
  journal_format = c("Taxon", "Systematic Botany", "Willdenowia")
)
}
\arguments{
\item{df}{A data frame containing a \code{scientificName} column or the individual components (\code{genus}, \code{specificEpithet}, etc.).}

\item{journal_format}{Character string specifying the journal style: \code{"Taxon"}, \code{"Systematic Botany"}, or \code{"Willdenowia"}.}
}
\value{
A data frame with the \code{scientificName} column standardized with italic markers.
}
\description{
Auxiliary function to formats genus, specific epithets, and infraspecific epithets names in italics, using Markdown syntax (*).
}
