% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_nomenclature.R
\name{format_synonym}
\alias{format_synonym}
\title{Format Synonyms}
\usage{
format_synonym(
  df = NULL,
  group_data,
  accepted_name,
  journal_format = c("Taxon", "Systematic Botany", "Willdenowia"),
  include_types = TRUE
)
}
\arguments{
\item{df}{A data frame containing taxonomic columns (e.g., \code{acceptedScientificName}, \code{scientificName}, \code{basionym}, \code{namePublishedIn}, etc.).}

\item{group_data}{Data frame with raw data for names linked to an accepted name.}

\item{accepted_name}{String containing the accepted scientific name.}

\item{journal_format}{Character string specifying the journal style: \code{"Taxon"}, \code{"Systematic Botany"}, or \code{"Willdenowia"}.}

\item{include_types}{Logical. If \code{"TRUE"}, looks for and formats type specimens associated with the synonym.}
}
\value{
A formatted character string containing the synonymy block.
}
\description{
Organize and format the list of homotypic and heterotypic synonyms for an correct name.
}
