% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_template.R
\name{get_template}
\alias{get_template}
\title{Download a template CSV for headeR}
\usage{
get_template(dir = "data_template", file_name = "template_headeR.csv")
}
\arguments{
\item{dir}{Character string defining the directory where the file will be saved. Default is \code{"results"}.}

\item{file_name}{Character string defining the name of the resulting Markdown file. Default is \code{"header_complete.md"}.}
}
\description{
Download a template CSV for headeR
}
