% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heading_data}
\alias{heading_data}
\title{Example dataset for headeR}
\format{
A data frame with 142 rows and 51 variables:
\describe{
\item{scientificName}{The scientific name of taxon with authorship information (Darwin Core column).}
\item{status}{Taxonomic status (e.g., 'correct name', 'synonym', 'nom. inval.', 'nom. illeg.').}
\item{namePublishedIn}{A reference for the publication in which the scientificName was originally established under the rules of the Madrid Code, following International Plant Names Index (Darwin Core column).}
\item{series}{The series of the publication (if applicable).}
\item{volume}{The volume of the publication (if applicable).}
\item{suppl}{The supplement of the publication (if applicable).}
\item{issue}{The issue or number of the publication (if applicable).}
\item{page}{The specific page where the name was described}
\item{namePublishedInYear}{The four-digit year in which the scientificName was published (including effective publication details, e.g., '1986 publ. 1987', following Plants of the World model) (Darwin Core column).}
\item{basisOfRecord}{The nature of the data record (e.g., PreservedSpecimen).}
\item{illustrationSource}{Source of the illustration for the taxon.}
\item{digitalObject}{Link or identifier for the digital representation of the specimen.}
\item{illustrationPublishedIn}{Reference where the illustration was published.}
\item{collectionCode}{The acronym identifying the herbarium institution from which the record was derived (e.g., UB, K, P, CEN, according with Thiers (2025), continuously updated) (Darwin Core column).}
\item{barcode}{The unique catalog number or barcode of the specimen.}
\item{recordedBy}{The name(s) of the responsible for recording the original Occurence. The one who applies a personal identifier (recordNumber) should be listed first. (Darwin Core column)}
\item{recordNumber}{An identifier given to the Occurrence at the time it was recorded. (Darwin Core column)}
\item{day}{The integer day of the month on which the Event occurred. (Darwin Core column)}
\item{inferredDay}{Inferred day of collection when missing from label.}
\item{month}{Month of collection (as spelled in the source dataset; e.g. "Jan.", "Feb.", "Mar.", "Apr.") (Darwin Core column).}
\item{inferredMonth}{Inferred month of collection.}
\item{year}{The four-digit year in which the Event occurred, according to the Common Era Calendar (Darwin Core column).}
\item{inferredYear}{Inferred year of collection.}
\item{phenology}{Phenological state of the sample (e.g., fl., fr.).}
\item{country}{The name of the country or major administrative unit in which the Location occurs (Darwin Core column).}
\item{inferredCountry}{Inferred country when missing from the original label.}
\item{stateProvince}{The name of the next smaller administrative region than country (state, province, canton, department, region, etc.) in which the Location occurs (Darwin Core column).}
\item{inferredStateProvince}{Inferred stateProvince when missing from the original label.}
\item{municipality}{The full, unabbreviated name of the next smaller administrative region than county (city, municipality, etc.) in which the Location occurs (Darwin Core column).}
\item{inferredMunicipality}{Inferred municipality when missing from the original label.}
\item{locality}{The specific description of the place (Darwin Core column).}
\item{inferredLocality}{Inferred locality when missing from the original label.}
\item{latitude}{Latitude coordinate (dwc:verbatimLatitude).}
\item{longitude}{Longitude coordinate (dwc:verbatimLongitude).}
\item{altitude}{Elevation of the collection in meters (dwc:verbatimElevation).}
\item{sheetNum}{Number of sheets in the Herbarium specimen.}
\item{category}{Specimen type category (e.g., holotype, isotype, paratype).}
\item{view}{Indication of type specimen viewing/verification (e.g., !).}
\item{viewPhoto}{Indication of type specimen photo viewing/verification (e.g., !).}
\item{typeDesignated}{Whether the type was designated by someone (by) or whether it will be designated in the work carried out (here).}
\item{typeAuthorship}{Authorship of the type designation.}
\item{typePublishedIn}{Publication where the type was designated, following International Plant Names Index.}
\item{typeSeries}{Series of the type designation publication.}
\item{typeVolume}{Volume of the type designation publication.}
\item{typeSuppl}{Supplement of the type designation publication.}
\item{typeIssue}{Issue of the type designation publication.}
\item{typePage}{Page of the type designation publication.}
\item{typePublishedInYear}{Year of the type designation publication.}
\item{basionym}{The basionym (botany) of the scientificName. The legitimate, previously published name on which a new combination or name at new rank is based. The basionym does not itself have a basionym; it provides the final epithet, name, or stem of the new combination or name at new rank (Art. 6.10) (Darwin Core column).}
\item{acceptedScientificName}{The currently accepted scientific name for the scientificName (Darwin Core column).}
\item{notes}{Additional taxonomic, nomenclatural, or specimen remarks.}
}
}
\source{
Data compiled by Lara Serpa Jaegge Deccache from herbaria and taxonomic literature for the project "Systematics, Evolution and Biogeography of \emph{Ouratea} Aubl. (Ochnaceae)".
}
\usage{
heading_data
}
\description{
A comprehensive dataset containing nomenclatural and type specimen data of Ochnaceae. This dataset was designed to test the formatting of taxonomic headings (following the Madrid Code 2025; Turland et al. 2025). The data follows most of the Darwin Core standards.
}
\details{
IMPORTANT: This data structure is the required model for successful processing
with \code{header_simple} and \code{header_complete}.
}
\keyword{datasets}
