\name{zigguserHeader-package}
\alias{zigguserHeader-package}
\alias{zigguserHeader}
\docType{package}
\title{\packageTitle{zigguserHeader}}
\description{\packageDescription{zigguserHeader}}
\section{Package Content}{\packageIndices{zigguserHeader}}
\author{\packageAuthor{zigguserHeader}}
\section{Maintainer}{\packageMaintainer{zigguserHeader}}
\keyword{package}
